backend/requirements.txt
# Django 博客后端依赖包

# ============================================
# 核心框架
# ============================================
Django==5.2.9
djangorestframework==3.16.0
django-cors-headers==4.9.0
django-filter==25.2

# ============================================
# 认证
# ============================================
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# ============================================
# 数据库
# ============================================
mysqlclient==2.2.7
PyMySQL==1.1.1
pymysql[rsa]==1.1.1

# ============================================
# Redis
# ============================================
redis==5.2.1
django-redis==5.4.0

# ============================================
# Elasticsearch
# ============================================
elasticsearch==8.17.2
django-elasticsearch-dsl==8.0

# ============================================
# Celery 异步任务
# ============================================
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.6.0
kombu==5.4.2

# ============================================
# API 文档
# ============================================
drf-yasg==1.21.8
setuptools==75.8.0
uritemplate==4.1.1

# ============================================
# 内容处理
# ============================================
markdown==3.7
Pillow==11.1.0
pygments==2.19.1

# ============================================
# 环境变量配置
# ============================================
python-decouple==3.8
python-dotenv==1.0.1

# ============================================
# 工具库
# ============================================
python-dateutil==2.9.0.post0
pytz==2024.2

# ============================================
# 开发工具 (可选)
# ============================================
django-debug-toolbar==4.4.6
ipython==8.31.0
black==25.1.0
flake8==7.1.1

# ============================================
# 测试
# ============================================
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==34.0.0

# ============================================
# 性能优化 (可选)
# ============================================
django-cache-url==3.4.5

# ============================================
# 安全
# ============================================
cryptography==45.0.0

# ============================================
# 监控和日志
# ============================================
sentry-sdk==2.20.0
